+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## Unreleased
+### Fix
+ - Configuration
+  - fix "Update FieldRoles" does not map some of the field roles in NG format to 7.3 which is used by the Web Components
+
 ## [v4.0.0-rc.3] - 2022.05.13
 ### Add
  - Export